requirements.txt
pyyaml==6.0
watchdog==2.1.9
cryptography==38.0.4
flask==2.2.3
werkzeug==2.2.3
requests==2.28.2
pyinstaller==6.10.0
PySimpleGUI>=5.0.8; platform_system=="Linux"
PySimpleGUI==4.60.5; platform_system=="Windows"
python-dotenv==1.0.0
pytest==7.3.1 